Improve speedcode
By using the X-position LSB/MSB of the pixels in the plot routine and
storing them in the clear routine, we save 22 cycles per pixel in the
clear routine and add 8 cycles per pixel in the plot routine. Resulting
in a net saving of 22 - 8 = 14 cycles per pixel. Which means a total
saving of 14 * 64 = 896 cycles.